cyw43_ctrl.c: Use port-defined extint_set and GPIO IT constants.
The `extint_set` function and `GPIO_IT_` constants are only defined
for the `stm32` port. To improve the portability of the host-wake/OOB
mode, those names are replaced with `CYW43_` constants to be defined
in ports `cyw43_configport.h`files, when they're used.

NOTE this is a breaking change which will require updating at least
the `stm32/cyw43_configport.h` file.