src/cyw43_ll: Fix issue with alignment of NVRAM download.
Prior to this commit, it's possible for the NVRAM size to be aligned twice:
once to 64 bytes before the call to cyw43_download_resource(), then again
in that function to 256 bytes.  If the 256 byte alignment increases the
length then the following assertion will fail:

    assert((dest_addr&BACKPLANE_ADDR_MASK)+sz) <= (BACKPLANE_ADDR_MASK+1)

This commit fixes that by only aligning to 64, which is all that's required
for SDIO bus transfers (and aligns to 4 for SPI buses).

Signed-off-by: Damien George <dpgeorge@georgerobotics.com.au>