src/cyw43_spi: Add support for gSPI transport.
This adds support for (general) SPI as a transport backend to communicate
with the cyw43xx chip.  To use this transport, a port must provide the
`cyw43_spi_xxx` functions defined in `cyw43_spi.h`, and enable the
configuration option `CYW43_USE_SPI`.  The interrupt pin is configured as
`CYW43_PIN_WL_IRQ`.

Signed-off-by: Damien George <dpgeorge@georgerobotics.com.au>